% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_labkey.R
\name{board_labkey}
\alias{board_labkey}
\title{Use a LabKey folder as a board}
\usage{
board_labkey(
  board_alias = NULL,
  base_url,
  folder,
  subdir = "pins",
  versioned = TRUE,
  api_key = Sys.getenv("LABKEY_API_KEY"),
  cache = NULL
)
}
\arguments{
\item{base_url}{Url of Labkey server}

\item{folder}{Folder within this server that this board will occupy}

\item{api_key}{API key to use for LabKey authentication. If not specified, will use \code{LABKEY_API_KEY}}
}
\description{
Pin data to a folder on a LabKey server
}
\details{
\code{board_labkey()} is powered by the Rlabkey package \url{https://github.com/cran/Rlabkey}
}
\examples{
\dontrun{
board <- board_labkey("pins-test-labkey")
board \%>\% pin_write(mtcars)
board \%>\% pin_read("mtcars")
}
}
