% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_labkey.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{pin_read}
\alias{pin_write}
\alias{pin_versions}
\alias{pin_list}
\alias{pin_exists}
\alias{pin_version_delete}
\alias{pin_delete}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{pins}{\code{\link[pins]{pin_delete}}, \code{\link[pins]{pin_exists}}, \code{\link[pins]{pin_list}}, \code{\link[pins]{pin_read}}, \code{\link[pins:pin_versions]{pin_version_delete}}, \code{\link[pins]{pin_versions}}, \code{\link[pins:pin_read]{pin_write}}}
}}

