% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_labkey.R
\name{board_labkey}
\alias{board_labkey}
\title{Use a LabKey folder as a board}
\usage{
board_labkey(
  board_alias = NULL,
  base_url,
  folder,
  subdir = "pins",
  versioned = TRUE,
  api_key = Sys.getenv("LABKEY_API_KEY"),
  cache = NULL
)
}
\arguments{
\item{board_alias}{alias of the board to be used for cache storage}

\item{base_url}{baseUrl of Labkey server}

\item{folder}{folderPath; folder path within the LabKey server to read/write pins}

\item{subdir}{subdirectory within the LabKey folder (aka remoteFilePath) where pin should be stored (default "pins")}

\item{versioned}{T or F whether to version the pin (default T)}

\item{api_key}{API key to use for LabKey authentication. If not specified, will use \code{LABKEY_API_KEY}}

\item{cache}{where to store board cache (if not specified will use default pins cache location)}
}
\description{
Pin data to a folder on a LabKey server
}
\details{
\code{board_labkey()} is powered by the Rlabkey package \url{https://github.com/cran/Rlabkey}
}
\examples{
\dontrun{
board <- board_labkey("pins-test-labkey")
board \%>\% pin_write(mtcars)
board \%>\% pin_read("mtcars")
}
}
