% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinsLabkey-package.R
\docType{package}
\name{pinsLabkey-package}
\alias{pinsLabkey}
\alias{pinsLabkey-package}
\title{pinsLabkey: Extend Pins Package for 'LabKey' Boards}
\description{
Read and write datasets and other R objects to LabKey servers \url{https://www.labkey.com/}. Extends pins using 'Rlabkey' to enable pinning of data to LabKey "boards", with flexibility to specify which LabKey server and folder to pin to.
}
\author{
\strong{Maintainer}: Clara Amorosi \email{clara.amorosi@bms.com} (\href{https://orcid.org/0000-0002-0730-2576}{ORCID})

}
\keyword{internal}
