% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinsLabkey-package.R
\docType{package}
\name{pinsLabkey-package}
\alias{pinsLabkey}
\alias{pinsLabkey-package}
\title{pinsLabkey: Extends Pins for Labkey Boards}
\description{
Package to extend R pins for LabKey boards
}
\author{
\strong{Maintainer}: Clara Amorosi (\href{https://orcid.org/0000-0002-0730-2576}{ORCID})

}
\keyword{internal}
